govern Viewer attraction fixture structure
