+++ crates/oasis7_viewer/software_safe_src/gameplay_attraction_scenario.js
@@ -1,278 +1,12 @@
-export const TASK_GAME_076_SCENARIO_VERSION = "task-game-076.v1";
+import {
+  ATTRACTION_THRESHOLDS,
+  BASE_MODEL,
+  BASE_PLAYER_GAMEPLAY,
+  TASK_GAME_076_BEATS,
+  TASK_GAME_076_SCENARIO_VERSION,
+} from "./gameplay_attraction_fixture_data.js";
 
-const ATTRACTION_THRESHOLDS = Object.freeze({
-  requiredCards: 3,
-  minAverageHookScore: 4,
-  minAverageReplayIntent: 4,
-  minMeaningfulDecisionCount: 2,
-  minRewardOrUnlockCount: 1,
-  maxStallOrWaitPeriods: 1,
-  minEffectivePlayMinutes: 30,
-  minPlayerOperationCount: 18,
-  minContentUnitCount: 8,
-  minDistinctActionFamilyCount: 6,
-  maxPassiveWaitShare: 0.25,
-});
-
-const BASE_MODEL = Object.freeze({
-  agents: {
-    "agent-0": {
-      id: "agent-0",
-      name: "Agent 0",
-      location_id: "loc-0",
-      pos: { x_cm: 2_900_000, y_cm: 3_450_000, z_cm: 0 },
-      resources: { alloy: 3 },
-    },
-  },
-  locations: {
-    "loc-0": {
-      id: "loc-0",
-      name: "Factory Anchor",
-      pos: { x_cm: 7_150_000, y_cm: 2_200_000, z_cm: 0 },
-      profile: { radius_cm: 55_000, radiation_emission_per_tick: 0, material: "silicate" },
-      resources: { iron: 0 },
-    },
-  },
-  agent_prompt_profiles: {},
-  agent_execution_debug_contexts: {
-    "agent-0": {
-      provider_mode: "deterministic_mock",
-      execution_mode: "phase_1",
-      environment_class: "task_game_076_scenario",
-      observation_schema_version: "viewer.v1",
-      action_schema_version: "gameplay_action.v1",
-      provider_check_status: "ok",
-      provider_check_source: "scenario_driver",
-      fallback_reason: null,
-      provider_reported_capabilities: ["gameplay_action", "agent_chat"],
-      provider_reported_supported_action_sets: ["gameplay_action", "agent_chat"],
-    },
-  },
-});
-
-const BASE_PLAYER_GAMEPLAY = Object.freeze({
-  stage_id: "post_onboarding",
-  stage_status: "blocked",
-  execution_state: "blocked",
-  accepted_intent_id: "gameplay_action:build_factory_smelter_mk1",
-  intent_summary: "Queue build_factory_smelter_mk1 for agent-0",
-  intent_scope: "gameplay_action",
-  intent_target: "agent-0",
-  status_reason: "Advance 1-2 steps to apply the queued build action.",
-  last_world_change: "Smelter build request reached factory-0; iron shortage blocks construction.",
-  resume_anchor: "Recover the first sustainable line",
-  resume_next_step: "Advance after replenishing materials to confirm the line resumes.",
-  goal_id: "post_onboarding.recover_capability",
-  goal_kind: "RecoverCapability",
-  goal_title: "Recover sustainable capability",
-  objective: "Stabilize the first production line before expanding.",
-  progress_detail: "The primary line is blocked by missing material input.",
-  progress_percent: 68,
-  blocker_kind: "material_shortage",
-  blocker_detail: "iron input exhausted at factory-0",
-  causality_kind: "world_constraint",
-  causality_detail: "iron input exhausted at factory-0",
-  next_step_hint: "Replenish upstream materials, then advance again to confirm the line resumes.",
-  branch_hint: null,
-  can_interrupt: true,
-  can_reprioritize: true,
-  replacement_intent_summary: "Switch from stalled smelter build to material recovery.",
-  handoff_result: "Old build intent stays paused until recovery proves material input.",
-  first_win_goal_id: "small_player.first_industrial_win",
-  player_action: "queued build_factory_smelter_mk1 and inspected the material blocker",
-  world_change_due_to_player: "factory-0 exposed a recoverable iron shortage instead of silent waiting",
-  player_leverage_verdict: "watch: recovery can restore the first capability",
-  leverage_class: "repair_elasticity",
-  same_loop_repeat_count: 2,
-  grind_only_flag: false,
-  major_power_dependency_status: "independent_path_available",
-  repair_available: true,
-  rebuild_available: true,
-  pivot_available: true,
-  recovery_path_kind: "repair_rebuild_or_pivot",
-  recovery_path_detail: "replenish iron, rebuild the local line, or pivot to a lower-input branch",
-  available_actions: [
-    {
-      action_id: "build_factory_smelter_mk1",
-      target_agent_id: "agent-0",
-      label: "Build smelter mk1",
-      protocol_action: "gameplay_action.submit",
-      disabled_reason: null,
-    },
-    {
-      action_id: "advance_step",
-      label: "Advance 1 step",
-      protocol_action: "live_control.step",
-      target_agent_id: null,
-      disabled_reason: null,
-    },
-    {
-      action_id: "request_snapshot",
-      label: "Request snapshot",
-      protocol_action: "world.request_snapshot",
-      target_agent_id: null,
-      disabled_reason: null,
-    },
-  ],
-  recent_feedback: {
-    action: "step",
-    stage: "blocked",
-    effect: "gameplay blocked before requested advance completed: logicalTime +0, eventSeq +0",
-    reason: "latest live control was blocked before runtime advance",
-    hint: "inspect the blocker details, recover the line, then advance again to confirm progress.",
-    delta_logical_time: 0,
-    delta_event_seq: 0,
-  },
-  agent_claim: null,
-});
-
-export const TASK_GAME_076_BEATS = Object.freeze([
-  {
-    time: "0-1m",
-    beat: "identity_goal_landing",
-    requiredCommands: ["software_safe_ui", "live_browser_30m_playthrough", "live_browser_30m_ui_click_playthrough",
-      "live_aw_test_completeness_playthrough"],
-    liveOptional: true,
-    assertion: "identity/goal/blocker/next are visible and diagnostics are demoted",
-    gapStatus: "goal_clarity_unverified",
-  },
-  {
-    time: "1-3m",
-    beat: "first_control_proof",
-    requiredCommands: [
-      "viewer_semantic_contract",
-      "runtime_control_feeling",
-      "live_browser_player_path",
-      "live_browser_30m_playthrough",
-      "live_browser_30m_ui_click_playthrough",
-      "live_aw_test_completeness_playthrough",
-    ],
-    liveOptional: true,
-    assertion: "play/step/action changes accepted intent, execution state, and Control Proof",
-    gapStatus: "first_control_success_unverified",
-  },
-  {
-    time: "3-5m",
-    beat: "first_consequence_read",
-    requiredCommands: [
-      "viewer_semantic_contract",
-      "runtime_control_feeling",
-      "live_pure_api_gameplay",
-      "live_browser_30m_playthrough",
-      "live_browser_30m_ui_click_playthrough",
-      "live_aw_test_completeness_playthrough",
-    ],
-    liveOptional: true,
-    assertion: "player_action maps to world_change_due_to_player and What I caused is visible",
-    gapStatus: "world_activity_only_risk",
-  },
-  {
-    time: "5-7m",
-    beat: "first_choice_hook",
-    requiredCommands: [
-      "viewer_semantic_contract",
-      "runtime_control_feeling",
-      "live_browser_player_path",
-      "live_browser_30m_playthrough",
-      "live_browser_30m_ui_click_playthrough",
-      "live_aw_test_completeness_playthrough",
-    ],
-    liveOptional: true,
-    assertion: "recommended action has label, cost/blocker/expected result, executable or disabled reason",
-    gapStatus: "meaningful_decision_unverified",
-  },
-  {
-    time: "7-10m",
-    beat: "first_blocker_recovery",
-    requiredCommands: [
-      "runtime_no_progress_recovery",
-      "runtime_chain_sync_blocker",
-      "software_safe_ui",
-      "live_browser_30m_playthrough",
-      "live_browser_30m_ui_click_playthrough",
-      "live_aw_test_completeness_playthrough",
-    ],
-    liveOptional: true,
-    assertion: "blocker taxonomy, recovery CTA, and next step exist; blocked is not collapsed into timeout",
-    gapStatus: "recovery_path_unverified",
-  },
-  {
-    time: "8-12m",
-    beat: "first_visible_output",
-    requiredCommands: [
-      "viewer_semantic_contract",
-      "live_pure_api_gameplay",
-      "live_browser_30m_playthrough",
-      "live_browser_30m_ui_click_playthrough",
-      "live_aw_test_completeness_playthrough",
-    ],
-    liveOptional: true,
-    assertion: "first output exposes input/output/new use or next value",
-    gapStatus: "reward_or_unlock_unverified",
-  },
-  {
-    time: "12-18m",
-    beat: "persistent_capability",
-    requiredCommands: [
-      "runtime_persist_backfill",
-      "live_pure_api_gameplay",
-      "live_browser_30m_playthrough",
-      "live_browser_30m_ui_click_playthrough",
-      "live_aw_test_completeness_playthrough",
-    ],
-    liveOptional: true,
-    assertion: "later snapshot still exposes progress/factory/recipe/recovery state",
-    gapStatus: "capability_persistence_unverified",
-  },
-  {
-    time: "18-23m",
-    beat: "first_expansion_tradeoff",
-    requiredCommands: [
-      "viewer_semantic_contract",
-      "live_pure_api_gameplay",
-      "live_browser_30m_playthrough",
-      "live_browser_30m_ui_click_playthrough",
-      "live_aw_test_completeness_playthrough",
-    ],
-    liveOptional: true,
-    assertion: "branch/choice appears and does not reduce to repeated grind",
-    gapStatus: "branch_offer_unverified",
-  },
-  {
-    time: "20-25m",
-    beat: "agency_correction",
-    requiredCommands: [
-      "viewer_semantic_contract",
-      "software_safe_ui",
-      "attraction_sufficiency_cards",
-      "live_pure_api_gameplay",
-      "live_browser_30m_playthrough",
-      "live_browser_30m_ui_click_playthrough",
-      "live_aw_test_completeness_playthrough",
-    ],
-    liveOptional: true,
-    assertion: "Agency Moves exposes interrupt/reprioritize/correction/handoff truth or unavailable reason",
-    gapStatus: "agency_move_unverified",
-  },
-  {
-    time: "25-30m",
-    beat: "return_hook_small_player_value",
-    requiredCommands: [
-      "viewer_semantic_contract",
-      "software_safe_ui",
-      "bevy_visual_probe",
-      "attraction_sufficiency_cards",
-      "live_pure_api_gameplay",
-      "live_browser_30m_playthrough",
-      "live_browser_30m_ui_click_playthrough",
-      "live_aw_test_completeness_playthrough",
-    ],
-    liveOptional: true,
-    visualOptional: true,
-    assertion: "First Win, Mature-World Continuation, Share Replay, leverage, and optional visual hierarchy are readable",
-    gapStatus: "return_hook_unverified",
-  },
-]);
+export { TASK_GAME_076_BEATS, TASK_GAME_076_SCENARIO_VERSION } from "./gameplay_attraction_fixture_data.js";
 
 const CONTENT_VOLUME_SUPPLEMENT_SEGMENTS = Object.freeze([
   "diagnosis_focus",
